Move testing logic out of executable
