+++ exe/cp-regression
@@ -6,8 +6,7 @@ if $PROGRAM_NAME == __FILE__
   require 'bundler/setup'
 end
 
-require 'regression'
-require 'yaml'
+require 'cocoapods-regression'
 
 if ENV['CP_REGRESSION'].nil?
   $stderr.puts 'Please provide path to development CP in CP_REGRESSION env var.'
@@ -16,45 +15,8 @@ end
 
 if ARGV[0].nil?
   puts 'Available regression tests:'
-
-  Dir.foreach('apps') do |file|
-    next unless file.end_with?('.yaml')
-    puts "\t#{File.basename(file, '.yaml')}"
-  end
-
+  puts CocoaPods::Regression.apps.map { |app| "\t#{app}" }
   exit(0)
 end
 
-app_config_path = "apps/#{ARGV[0]}.yaml"
-
-unless File.exist?(app_config_path)
-  $stderr.puts "Regression test `#{ARGV[0]}` does not exist."
-  exit(1)
-end
-
-app = YAML.load(File.new(app_config_path))
-checkout = CocoaPods::Regression.clone(app['source'], ARGV[0])
-xcode = CocoaPods::Regression::Xcode.new(app)
-
-builder = CocoaPods::Regression::Builder.new(app)
-builder.working_directory = checkout
-
-# Build using current stable CP
-unless builder.build(xcode)
-  $stderr.puts "Couldn't build using stable CP, aborting."
-  exit(3)
-end
-
-# Build by using new CP only
-unless builder.build(xcode, ENV['CP_REGRESSION'])
-  $stderr.puts "Couldn't do a clean build using development CP, aborting."
-  exit(3)
-end
-
-# Build by using stable CP, then new CP
-unless builder.build_upgrade(xcode, ENV['CP_REGRESSION'])
-  $stderr.puts "Couldn't do an upgrade build using development CP, aborting."
-  exit(3)
-end
-
-puts 'The development CP is 👍'
+CocoaPods::Regression.test!(ARGV[0], ENV['CP_REGRESSION'])
