+++ lib/cocoapods-regression/cocoapods.rb
