+++ lib/cocoapods-regression/version.rb
