+++ lib/cocoapods-regression/xcode.rb
