+++ spec/spec_helper.rb
@@ -10,4 +10,4 @@ require 'bundler/setup'
 require 'bacon'
 require 'mocha-on-bacon'
 require 'pretty_bacon'
-require 'regression'
+require 'cocoapods-regression'
